Added checks before writes and membank setting
The related issue is Issue #10 in lis2dux12-pid repository (see
https://github.com/STMicroelectronics/lis2dux12-pid/issues/10).

During updates, namely a read and a write on the same content,
there were no checks to guarantee that reads have retrived the data,
and the writes could use random content when this issue happens.
The same fix is also applied when a pointer is changed based on the
result of a read.

Another set of analogous checks was added to the membank so it jumps
to the exit condition (resetting the membank to its correct state)
before returning the error.

Added comments and rewrite of ln_pg_write/reads

Signed-off-by: Gioele Fiorenza <gioele.fiorenza2000@gmail.com>